# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "cron-as-a-service",
    "description": "Cron As A Service",
    "version": "0.9.0"
  },
  "servers": [
    {
      "url": "https://www.cronasaservice.com/api"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "List possible API calls",
        "description": "List api calls",
        "operationId": "Api.ListCalls",
        "parameters": [
          {
            "name": "apikey",
            "description": "Api Key.",
            "in": "query",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "method": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "method",
                    "path"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/jobs": {
      "post": {
        "summary": "Create a new job",
        "description": "Create a new job in a project",
        "operationId": "Jobs.CreateJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apikey": {
                    "type": "string"
                  },
                  "expression": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  }
                },
                "required": [
                  "apikey",
                  "expression",
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expression": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "expression",
                    "url"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all Jobs for a project (you must be an authorized user of the project)",
        "description": "Get all jobs for a project",
        "operationId": "Jobs.GetProjectJobs",
        "parameters": [
          {
            "name": "apikey",
            "description": "Api Key.",
            "in": "query",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "id",
            "description": "ID of the project.",
            "in": "path",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expression": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "expression",
                    "url"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "post": {
        "summary": "Create a new job",
        "description": "Creates a new job",
        "operationId": "Jobs.CreateJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apikey": {
                    "type": "string"
                  },
                  "expression": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  }
                },
                "required": [
                  "apikey",
                  "expression",
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expression": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "expression",
                    "url"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all Jobs for the authenticated user",
        "description": "Get all Jobs for the authenticated user",
        "operationId": "Jobs.GetJobs",
        "parameters": [
          {
            "name": "apikey",
            "description": "Api Key.",
            "in": "query",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expression": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "expression",
                    "url"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}/alarms": {
      "post": {
        "summary": "Add an alarm to an existing job",
        "description": "Add an alarm to a job",
        "operationId": "Jobs.EditJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "statusCode": {
                    "type": "number"
                  },
                  "jsonPath": {
                    "type": "string"
                  },
                  "jsonPathResult": {
                    "type": "string"
                  }
                },
                "required": [
                  "statusCode",
                  "jsonPath",
                  "jsonPathResult"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "jsonPath": {
                      "type": "string"
                    },
                    "jsonPathResult": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "jsonPath",
                    "jsonPathResult"
                  ]
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {},
                      "required": []
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}